[refactor] CustomView adapter 파일 패키징
